C6 Exchange_tab set up, Coins collectable

+++ Coinz/app/src/main/java/com/example/kelbel/frederik/coinz/NastyCoin.kt
@@ -5,5 +5,5 @@ data class NastyCoin (
         var value: Float,
         var currency: String,
         var marker_symbol: String,
-        var coordinates: Pair<Float, Float>
+        var coordinates: Pair<Double, Double>
 )
